Modulo14_DRF_API/requirements/dev_requirements.txt
#  Essas são as dependencias para o modo desenvolvedor

-r common.txt

coverage==6.3.2
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2